+++ RecyclerView/app/src/main/java/net/boondockradio/recyclerview/StatesAdapter.java
@@ -1,13 +1,16 @@
 package net.boondockradio.recyclerview;
 
+import android.content.Context;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.TextView;
+import android.widget.Toast;
 
 import java.util.ArrayList;
 
-public class StatesAdapter extends RecyclerView.Adapter<ItemViewHolder> {
+public class StatesAdapter extends RecyclerView.Adapter<StatesAdapter.ItemViewHolder> {
 
     private ArrayList<String> states;
 
@@ -24,11 +27,25 @@ public class StatesAdapter extends RecyclerView.Adapter<ItemViewHolder> {
     }
 
     @Override
-    public void onBindViewHolder(ItemViewHolder holder, int position) {
-        final String data;
-        data = states.get(position);
+    public void onBindViewHolder(ItemViewHolder holder, final int position) {
+        String stateText;
+        stateText = states.get(position);
 
-        holder.textView.setText(data);
+        final Context context = holder.itemView.getContext();
+        holder.textView.setText(stateText);
+
+        holder.textView.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                int hateIndex = (position == states.size() - 1) ? 0 : position;
+                String message = context.getString(R.string.states_lover, states.get(position), states.get(hateIndex));
+                Toast.makeText(
+                        context,
+                        message,
+                        Toast.LENGTH_SHORT
+                ).show();
+            }
+        });
     }
 
     @Override
@@ -36,4 +53,12 @@ public class StatesAdapter extends RecyclerView.Adapter<ItemViewHolder> {
         return states.size();
     }
 
+    class ItemViewHolder extends RecyclerView.ViewHolder {
+        public TextView textView;
+
+        public ItemViewHolder(View v) {
+            super(v);
+            textView = (TextView) v.findViewById(R.id.text_raw_item);
+        }
+    }
 }
