+++ RecyclerView/app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
     <string name="app_name">RecyclerView</string>
 
-    <string name="states_lover">I love %1s</string>
+    <string name="states_lover">I love %1$s, not %2$s</string>
 </resources>
